requirements.txt
streamlit>=1.35.0
psycopg[binary]>=3.1.19
pgvector>=0.2.5
openai>=1.40.0
requests>=2.32.0
python-dateutil>=2.9.0
python-pptx>=0.6.23
pymupdf>=1.24.0
pillow>=10.3.0
pytesseract>=0.3.10
numpy>=1.26.0
